+++ Catch_serialization_tests/Catch_serialization_tests.cpp
@@ -100,14 +100,14 @@ TEST_CASE("simple serialzie", "[serialization_tests]") {
 
 		ptr->thingy_get_obj_value(target).push_back(3.0f);
 		ptr->thingy_get_obj_value(target).push_back(4.0f);
-		ptr->thingy_pooled_v_add_unique(target, -5i16);
+		ptr->thingy_pooled_v_add_unique(target, int16_t(-5));
 		ptr->thingy_resize_big_array(2);
 		ptr->thingy_resize_big_array_bf(4);
 
 		ptr->thingy_set_big_array(target, 0, 1.5f);
 		ptr->thingy_set_big_array(target, 1, 3.5f);
 		ptr->thingy_set_big_array_bf(target, 3, true);
-		target.get_i_value() = 35i16;
+		target.get_i_value() = int16_t(35);
 		target.get_f_value() = 7.5f;
 		target.pooled_v_push_back(3);
 		target.pooled_v_push_back(5);
@@ -182,14 +182,14 @@ TEST_CASE("serialzie complex props", "[serialization_tests]") {
 
 		ptr->thingy_get_obj_value(target).push_back(3.0f);
 		ptr->thingy_get_obj_value(target).push_back(4.0f);
-		ptr->thingy_pooled_v_add_unique(target, -5i16);
+		ptr->thingy_pooled_v_add_unique(target, int16_t(-5));
 		ptr->thingy_resize_big_array(2);
 		ptr->thingy_resize_big_array_bf(4);
 
 		ptr->thingy_set_big_array(target, 0, 1.5f);
 		ptr->thingy_set_big_array(target, 1, 3.5f);
 		ptr->thingy_set_big_array_bf(target, 3, true);
-		target.get_i_value() = 35i16;
+		target.get_i_value() = int16_t(35);
 		target.get_f_value() = 7.5f;
 		target.pooled_v_push_back(3);
 		target.pooled_v_push_back(5);
@@ -241,7 +241,7 @@ TEST_CASE("serialzie complex props", "[serialization_tests]") {
 		REQUIRE(target.get_pooled_v_at(2) == 5);
 		REQUIRE(ptr->thingy_get_obj_value(target).size() == 2);
 		REQUIRE(ptr->thingy_get_obj_value(target)[1] == 4.0f);
-		REQUIRE(ptr->thingy_pooled_v_contains(target, -5i16) == true);
+		REQUIRE(ptr->thingy_pooled_v_contains(target, int16_t(-5)) == true);
 		REQUIRE(ptr->thingy_get_big_array(target, 0) == 1.5f);
 		REQUIRE(ptr->thingy_get_big_array(target, 1) == 3.5f);
 		REQUIRE(ptr->thingy_get_big_array_bf(target, 3) == true);
@@ -264,7 +264,7 @@ TEST_CASE("deserialize a related object", "[serialization_tests]") {
 	ptr->create_thingy();
 
 
-	target.get_i_value() = 35i16;
+	target.get_i_value() = int16_t(35);
 	target.get_f_value() = 7.5f;
 
 	auto s1 = ptr->create_thingy2();
@@ -393,7 +393,7 @@ TEST_CASE("upgrading", "[serialization_tests]") {
 	ptr->create_thingy();
 	ptr->create_thingy();
 
-	target.get_i_value() = 35i16;
+	target.get_i_value() = int16_t(35);
 	target.get_f_value() = 7.5f;
 
 	auto s1 = ptr->create_thingy2();
